Moved ImageAdder to its own file
To make merging easier and such
